% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2tex.R
\name{lm2tex}
\alias{lm2tex}
\title{Convert lm model to latex syntax.}
\usage{
lm2tex(model, type, round_num)
}
\arguments{
\item{model}{The lm model that will be converted.}

\item{type}{"pop" or "fit" or "both" represents the model returned is either population model, fitted model or both.}

\item{round_num}{Number of significant digits.}
}
\value{
The relative latex syntax.
}
\description{
This function takes in the model generated by lm function and returns the latex syntax of either population model or fitted model or both.
}
\section{Functions}{
\itemize{
\item \code{lm2tex}: function for producing the latex syntax.
}}

