% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm2tex.R, R/lm2tex.R
\name{glm2tex}
\alias{glm2tex}
\alias{lm2tex}
\title{Convert lm model to latex syntax.}
\usage{
glm2tex(glm_model, glm_form, glm_type, glm_round_num)

lm2tex(model, type, round_num)
}
\arguments{
\item{glm_model}{The glm model that will be converted.}

\item{glm_form}{"log", "odds" or "prob" represents the model returned is either model in log form, model in terms of odds and model in terms of probabilities.}

\item{glm_type}{"pop", "fit" or "both" represents the model returned is either population model, fitted model or both.}

\item{glm_round_num}{Number of significant digits.}

\item{model}{The lm model that will be converted.}

\item{type}{"pop" or "fit" or "both" represents the model returned is either population model, fitted model or both.}

\item{round_num}{Number of significant digits.}
}
\value{
Print the relative latex syntax.

The relative latex syntax.
}
\description{
This function takes in the model generated by glm function and returns the latex syntax of either population model or fitted model or both.

This function takes in the model generated by lm function and returns the latex syntax of either population model or fitted model or both.
}
\section{Functions}{
\itemize{
\item \code{glm2tex}: function for producing the latex syntax of logistic regression model.

\item \code{lm2tex}: function for producing the latex syntax of the linear regression model.
}}

